removed redundancy in unit normals and reformatted
I removed a redundancy in the unit normal's in "normal_floats" where the same unit normal's were repeated. Because they are accessed with indices only a single copy of each unit normal vector is required.  I also added spaces in "normal_floats" to help the artist see the vectors within it (this could also be done to vert_floats if you think is helpful).
I then updated "indices" references to the new unit normal's indices and reformatted "indices" so that the triangles are grouped. This also makes it clear that the unit normals index follows each vertex index.