Make numpy an explicit install requirement
On a project I am working on, we are using `pip-compile` to generate the `requirements.txt` for an application. 

Currently this generates different results depending on whether or not `numpy` is installed on the system. This appears to be because of the conditional logic that only declares `numpy` as an install dependency if it's not already installed.

Shouldn't `numpy` just be part of the `install_requires` list? That is the old code is only stating the `numpy` is an explicit requirement if is not installed. 

If it is installed, wouldn't this make `numpy` an implicit install requirement?